pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "agentic-ai-bootcamp"
version = "0.1.0"
description = "AI Agentic Systems bootcamp project — CrewAI + Docker + Jenkins + Streamlit"
requires-python = ">=3.12"
dependencies = [
    "crewai>=0.80.0",
    "crewai-tools>=0.15.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "streamlit>=1.40.0",
    "python-dotenv>=1.0.0",
    "anthropic>=0.40.0",
    "requests>=2.32.0",
    "pydantic>=2.10.0",
    "pyyaml>=6.0.0",
    "pandas>=2.0.0",
    "pytest>=8.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
